+++ src/lib/random.c
@@ -14,7 +14,7 @@
 #include <openssl/rand.h>
 
 /**
- * Generates an unsigned type random number in range (inclusive lower and upper bounds)
+ * Generates an unsigned type random number in range (inclusive lower and exclusive upper bound)
  * @param lower lower boundary
  * @param upper upper boundary
  * @return random number
@@ -22,7 +22,7 @@
 unsigned randrange(unsigned lower, unsigned upper) {
     unsigned buf;
     if (RAND_bytes((unsigned char *)&buf, sizeof(buf)) == 1) {
-        return lower + buf / (UINT_MAX / (upper - lower + 1) + 1);
+        return lower + (unsigned) ((upper - lower) * (buf / (UINT_MAX + 1.0)));
     }
 
     MYMPD_LOG_ERROR(NULL, "Error generating random number in range");